feat: courses vers API :sparkles:
