+++ src/commons/services/course/Course.service.ts
@@ -1,5 +1,6 @@
-import CourseRepository from "./Course.repository";
+// import CourseRepository from "./Course.repository.mock";
 import {Course} from "../../models";
+import CourseRepository from "./Course.repository";
 
 /**
  * Service permettant de gérer les cours
@@ -25,17 +26,17 @@ class CourseService {
         throw new Error("Course not found");
     }
 
-    /**
-     * Récupère tous les cours ayant le label demandé
-     * @param label
-     */
-    async findCourseByLabel(label: string): Promise<Course[] | undefined> {
-        const courses = CourseRepository.findByLabel(label);
-        if (courses === undefined) {
-            throw new Error("Aucun cours trouvé avec ce label");
-        }
-        return Promise.resolve(courses);
-    }
+    // /**
+    //  * Récupère tous les cours ayant le label demandé
+    //  * @param label
+    //  */
+    // async findCourseByLabel(label: string): Promise<Course[] | undefined> {
+    //     const courses = CourseRepository.findByLabel(label);
+    //     if (courses === undefined) {
+    //         throw new Error("Aucun cours trouvé avec ce label");
+    //     }
+    //     return Promise.resolve(courses);
+    // }
 
     /**
      * Créé un cours ou le met à jour s'il existe déjà
@@ -45,13 +46,13 @@ class CourseService {
         return await CourseRepository.save(course);
     }
 
-    /**
-     * Ecrase le cours avec les nouvelles informations
-     * @param course
-     */
-    async replaceCourse(course: Course): Promise<Course> {
-        return await CourseRepository.replace(course);
-    }
+    // /**
+    //  * Ecrase le cours avec les nouvelles informations
+    //  * @param course
+    //  */
+    // async replaceCourse(course: Course): Promise<Course> {
+    //     return await CourseRepository.replace(course);
+    // }
 
     /**
      * Permet de rechercher un cours en fonction de son label et de sa date de début
